+++ src/thread/threads_posix.h
@@ -41,7 +41,7 @@ TEST_RESULT test() {
 }
 
 void *thread_function(void *dummyPtr) {
-  printf("Thread number %ld\n", pthread_self());
+  printf("Thread number %ld\n", static_cast<long int>(pthread_self()));
   pthread_mutex_lock(&mutex1);
   counter++;
   pthread_mutex_unlock(&mutex1);
